for lab

+++ pipe.h
@@ -11,5 +11,5 @@ World worldFromString(std::string);
 void castRays(World w);
 Intersection intersectRayWithSpheres(Ray, std::list<Sphere>, Camera);
 Intersection intersectRayWithPolygons(Ray, std::list<Polygon>, Camera);
-ublas::vector<double> calcPixelColor(Ray, ublas::vector<double>, ublas::vector<double>, Material, std::list<Light>, double, double);
+ublas::vector<float> calcPixelColor(Ray, ublas::vector<float>, ublas::vector<float>, Material, std::list<Light>, float, float);
 #endif /* PIPE_H */
